Games\Blackjack - Remove player parameter for keyboard
MonoGame has deprecated the support for multiple keyboards on the same PC, so we can remove all the code that used the deprecated feature.

Also made the out parameter PlayerIndex for keyboard default to 1 since it is never used in a meaningful way in the game as the game is not really multiplayer and removed some out parameters. A lot more thoughts and effort would be needed to acutally allow local multiplayer to work in this game.